refactor(broker): split version_check.rs into directory layout (C42 of PR-3)
Final commit of PR-3. Pure mechanical move. version_check.rs (607 lines)
becomes:

- version_check/mod.rs (~35 lines): module decls + pub(crate) use
  service::VersionCheckTask + pub use service::{format_version,
  pack_version, unpack_version, BROKER_VERSION}.
- version_check/error.rs (~36 lines): VersionCheckError enum +
  coded_error_impl!.
- version_check/service.rs (~587 lines): VersionCheckTask<P> struct +
  impl, BrokerService impl, the IVersionRegistry sol! ABI binding, the
  VERSION_REGISTRIES lookup table, lookup_registry / check_version
  helpers, the version-packing helpers (BROKER_VERSION, pack_version,
  unpack_version, format_version, const_str_to_u16), and the original
  test module.

External paths preserved:
- crate::version_check::VersionCheckTask (broker.rs)
- broker::version_check::{BROKER_VERSION, pack_version, unpack_version,
  format_version} (downstream-facing API surface; lib.rs's
  `pub mod version_check` declaration unchanged).
